refactor(tui): 画面ログを screen_log.py に切り出す
--debug-log 指定時だけ動くデバッグ機能で TUI の動作とは直交しており、
app.renderer._last_screen という prompt_toolkit の内部に触る唯一の箇所でも
あるため app.py から分ける。

フックの登録は attach_screen_log() にまとめ、app.py 側は 2 行にする。
