+++ src/redi/tui/__init__.py
@@ -1,4 +1,5 @@
-from redi.tui.app import dump_rendered_screen, run_issue_tui
+from redi.tui.app import run_issue_tui
+from redi.tui.screen_log import dump_rendered_screen
 from redi.tui.state import (
     IssueTabState,
     Renderable,
